# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AssetKeeper",
    "description": "AssetKeeper用API",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
...
  "paths": {
...
    "/accounts": {
      "get": {
        "tags": ["accounts"],
        "summary": "Get accounts",
        "description": "",
        "operationId": "getAccounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accounts"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Accounts"
                }
              }
            }
          },
          "404": {
            "description": "Account not found"
          }
        }
      },
      "post": {
        "tags": ["accounts"],
        "summary": "Create account",
        "description": "This can only be done by the logged in user.",
        "operationId": "createAccount",
        "requestBody": {
          "description": "Created account object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Accounts"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Accounts"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Accounts"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accounts"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Accounts"
                }
              }
            }
          }
        }
      }
    },
...
    "/accounts-with-balances": {
      "get": {
        "tags": ["accounts"],
        "summary": "Get accounts with balance history",
        "description": "",
        "operationId": "getAccountsWithBalances",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "2023-01-01"
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "2023-12-31"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                }
              }
            }
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Users": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "userName": {
            "type": "string",
            "example": "theUser"
          },
          "email": {
            "type": "string",
            "example": "john@email.com"
          },
          "password": {
            "type": "string",
            "example": "12345"
          },
          "role": {
            "type": "string",
            "example": "Admin"
          },
          "color": {
            "type": "string",
            "example": "#19C37D"
          }
        },
        "xml": {
          "name": "users"
        }
      },
      "Accounts": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "accountName": {
            "type": "string",
            "example": "Abank"
          },
          "accountType": {
            "type": "string",
            "example": "Bank"
          },
          "ownerId": {
            "type": "string",
            "example": "12345"
          }
        },
        "xml": {
          "name": "accounts"
        }
      },
...
    },
...
  }
}